rename appStore to reduxStore
