+++ src/index.js
@@ -3,11 +3,11 @@ import { render } from "react-dom";
 import { HashRouter } from "react-router-dom";
 import App from "./App";
 import "../less/styles.less";
-import store from "./stores/store";
+import reduxStore from "./stores/reduxStore";
 import { Provider } from "react-redux";
 
 render(
-  <Provider store={store}>
+  <Provider store={reduxStore}>
     <HashRouter>
       <App />
     </HashRouter>
